Move `eventsourcing.Sender` to `saga` package.
